extract host connector to trait
