+++ src/main/scala/org/broadinstitute/dsde/firecloud/dataaccess/HttpOntologyDAO.scala
@@ -5,7 +5,7 @@ import com.typesafe.scalalogging.slf4j.LazyLogging
 import org.broadinstitute.dsde.firecloud.FireCloudConfig
 import org.broadinstitute.dsde.firecloud.model.ModelJsonProtocol._
 import org.broadinstitute.dsde.firecloud.model.Ontology.TermResource
-import org.broadinstitute.dsde.firecloud.utils.RestJsonClient
+import org.broadinstitute.dsde.firecloud.utils.{HostConnector, RestJsonClient}
 import spray.http.Uri
 import spray.httpx.SprayJsonSupport._
 import spray.httpx.unmarshalling._
@@ -15,7 +15,7 @@ import scala.concurrent.{ExecutionContext, Future}
 
 
 class HttpOntologyDAO(implicit val system: ActorSystem, implicit val executionContext: ExecutionContext)
-  extends OntologyDAO with RestJsonClient with LazyLogging {
+  extends OntologyDAO with RestJsonClient with HostConnector with LazyLogging {
 
   override def search(term: String): Future[Option[List[TermResource]]] = {
     searchAsync(term)
