+++ src/main/scala/org/broadinstitute/dsde/firecloud/dataaccess/HttpRawlsDAO.scala
@@ -4,7 +4,7 @@ import akka.actor.ActorSystem
 import org.broadinstitute.dsde.firecloud.model.ErrorReportExtensions._
 import org.broadinstitute.dsde.firecloud.model.ModelJsonProtocol._
 import org.broadinstitute.dsde.firecloud.model._
-import org.broadinstitute.dsde.firecloud.utils.RestJsonClient
+import org.broadinstitute.dsde.firecloud.utils.{HostConnector, RestJsonClient}
 import org.broadinstitute.dsde.firecloud.{FireCloudConfig, FireCloudExceptionWithErrorReport}
 import org.broadinstitute.dsde.rawls.model.AttributeUpdateOperations._
 import org.broadinstitute.dsde.rawls.model.StatusJsonSupport._
@@ -26,7 +26,7 @@ import scala.util.control.NonFatal
   * Created by davidan on 9/23/16.
   */
 class HttpRawlsDAO( implicit val system: ActorSystem, implicit val executionContext: ExecutionContext )
-  extends RawlsDAO with RestJsonClient {
+  extends RawlsDAO with RestJsonClient with HostConnector {
 
   override def isRegistered(userInfo: UserInfo): Future[Boolean] = {
     userAuthedRequest(Get(rawlsUserRegistrationUrl))(userInfo) map { response =>
